+++ src/main/java/cherish/backend/member/dto/ChangeInfoRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ChangeInfoRequest {
-    private String memberEmail;
+    private String email;
     private String nickName; // 닉네임
     // 추가 정보
     private String jobName;
