+++ src/main/java/cherish/backend/member/dto/ChangePwdRequest.java
@@ -1,10 +1,9 @@
 package cherish.backend.member.dto;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 public class ChangePwdRequest {
-    private final String pwd;
+    private final String password;
     private final String email;
 }
