+++ src/main/java/cherish/backend/member/dto/MemberFormDto.java
@@ -1,8 +1,6 @@
 package cherish.backend.member.dto;
 
 import cherish.backend.member.model.enums.Gender;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import lombok.AllArgsConstructor;
 import lombok.Data;
